removeBadImages.sh: eliminate temporary file
* This command is normally called for a single file, and that file is normally good, so no output is generated.  Creating a blank output file is generating an unneeded disk write to the SD card.  Save output from a single file in a variable, and if there's a problem echo the variable.
* In the vast majority of the cases, this script doesn't write anything to disk.
* Fixed a shellcheck issue.

+++ scripts/removeBadImages.sh
@@ -83,9 +83,16 @@ else
 fi
 ERROR_WORDS="Huffman|Bogus|Corrupt|Invalid|Trunc|Missing|insufficient image data|no decode delegate|no images defined"
 
-# Save all output to a temp file so don't potentially swamp the system log file.
-OUTPUT="${ALLSKY_TMP}/removeBadImages.log"
-> ${OUTPUT}
+# Reduce writes to disk if possible.  This script is normally called once for each file,
+# and most files are good so no output is created and hence no reason to create a temporary
+# OUTPUT file.  Only use OUTPUT if we're doing a whole directory at once,
+# otherwise put a single file's output in a variable.
+if [ -n "${FILE}" ]; then		# looking at one file
+	OUTPUT=""
+else							# looking at a directory
+	OUTPUT="${ALLSKY_TMP}/removeBadImages.log"
+	> ${OUTPUT}
+fi
 
 typeset -i num_bad=0
 # If the low threshold is 0 it's disabled.
@@ -113,6 +120,7 @@ for f in ${IMAGE_FILES} ; do
 	else
 		# MEAN is a number between 0.0 and 1.0.
 		MEAN=$(${NICE} convert "${f}" -colorspace Gray -format "%[fx:image.mean]" info: 2>&1)
+		# shellcheck disable=SC2181
 		if [ $? -ne 0 ]; then
 			# Do NOT set BAD since this isn't necessarily a problem with the file.
 			echo -e "${RED}***${ME}: ERROR: 'convert ${f}' failed; leaving file.${NC}"
@@ -156,7 +164,11 @@ for f in ${IMAGE_FILES} ; do
 	fi
 
 	if [ "${BAD}" != "" ]; then
-		echo "${r} ${BAD}" >> "${OUTPUT}"
+		if [ -n "${FILE}" ]; then		# looking at one file, save message in variable
+			OUTPUT="${r} ${BAD}"
+		else							# looking at a directory, save message in tmp file
+			echo "${r} ${BAD}" >> "${OUTPUT}"
+		fi
 		[ ${DEBUG} = "false" ] && rm -f "${f}" "thumbnails/${f}"
 		let num_bad=num_bad+1
 	fi
@@ -164,15 +176,16 @@ done
 
 if [ $num_bad -eq 0 ]; then
 	# If only one file, "no news is good news".
-	[ "${FILE}" = "" ] && echo -e "\n${ME} ${GREEN}No bad files found.${NC}"
-	rm -f "${OUTPUT}"
+	if [ "${FILE}" = "" ]; then
+		echo -e "\n${ME} ${GREEN}No bad files found.${NC}"
+		rm -f "${OUTPUT}"
+	fi
 else
 	if [ "${FILE}" = "" ]; then
 		echo "${ME} ${num_bad} bad file(s) found and ${r}. See ${OUTPUT}."
 		# Do NOT remove ${OUTPUT} in case the user wants to look at it.
 	else	# only 1 file so show it
-		echo "${ME} File is bad: $(< "${OUTPUT}")"
-		rm -f "${OUTPUT}"
+		echo "${ME} File is bad: ${OUTPUT}"
 	fi
 fi
